requirements.txt
pandas
numpy
pygwalker
streamlit
openai
matplotlib
databutton
extra-streamlit-components
streamlit-option-menu
pillow
argparse